struct(T9): split UI input handlers out of app.rs
