+++ src/ui/input/mod.rs
@@ -1,0 +1,6 @@
+mod deals;
+mod file;
+mod main;
+mod marketing;
+mod pickers;
+mod setup;
