+++ src/ui/mod.rs
@@ -1,2 +1,3 @@
 pub mod app;
+mod input;
 mod render;
